# Add `#[inline]` to forwarding methods of `wasmparser::collections` types