--- a/crates/wasmparser/src/collections/set.rs
+++ b/crates/wasmparser/src/collections/set.rs
@@ -49,6 +49,7 @@ pub struct Set<T> {
 }
 
 impl<T> Default for Set<T> {
+    #[inline]
     fn default() -> Self {
         Self {
             inner: detail::SetImpl::default(),
@@ -58,6 +59,7 @@ impl<T> Default for Set<T> {
 
 impl<T> Set<T> {
     /// Clears the [`Set`], removing all elements.
+    #[inline]
     pub fn clear(&mut self) {
         self.inner.clear()
     }
@@ -66,6 +68,7 @@ impl<T> Set<T> {
     ///
     /// In other words, remove all elements `e` for which `f(&e)` returns `false`.
     /// The elements are visited in unsorted (and unspecified) order.
+    #[inline]
     pub fn retain<F>(&mut self, f: F)
     where
         T: Ord,
@@ -75,16 +78,19 @@ impl<T> Set<T> {
     }
 
     /// Returns the number of elements in the [`Set`].
+    #[inline]
     pub fn len(&self) -> usize {
         self.inner.len()
     }
 
     /// Returns `true` if the [`Set`] contains no elements.
+    #[inline]
     pub fn is_empty(&self) -> bool {
         self.inner.is_empty()
     }
 
     /// Returns an iterator that yields the items in the [`Set`].
+    #[inline]
     pub fn iter(&self) -> Iter<'_, T> {
         Iter {
             inner: self.inner.iter(),
@@ -97,6 +103,7 @@ where
     T: Eq + Hash + Ord,
 {
     /// Reserves capacity for at least `additional` more elements to be inserted in the [`Set`].
+    #[inline]
     pub fn reserve(&mut self, additional: usize) {
         #[cfg(not(feature = "no-hash-maps"))]
         self.inner.reserve(additional);
@@ -105,6 +112,7 @@ where
     }
 
     /// Returns true if the [`Set`] contains an element equal to the `value`.
+    #[inline]
     pub fn contains<Q>(&self, value: &Q) -> bool
     where
         T: Borrow<Q>,
@@ -114,6 +122,7 @@ where
     }
 
     /// Returns a reference to the element in the [`Set`], if any, that is equal to the `value`.
+    #[inline]
     pub fn get<Q>(&self, value: &Q) -> Option<&T>
     where
         T: Borrow<Q>,
@@ -128,6 +137,7 @@ where
     ///
     /// - Returns `true` if the set did not previously contain an equal value.
     /// - Returns `false` otherwise and the entry is not updated.
+    #[inline]
     pub fn insert(&mut self, value: T) -> bool {
         self.inner.insert(value)
     }
@@ -135,6 +145,7 @@ where
     /// If the set contains an element equal to the value, removes it from the [`Set`] and drops it.
     ///
     /// Returns `true` if such an element was present, otherwise `false`.
+    #[inline]
     pub fn remove<Q>(&mut self, value: &Q) -> bool
     where
         T: Borrow<Q>,
@@ -149,6 +160,7 @@ where
     /// The value may be any borrowed form of the set's element type,
     /// but the ordering on the borrowed form *must* match the
     /// ordering on the element type.
+    #[inline]
     pub fn take<Q>(&mut self, value: &Q) -> Option<T>
     where
         T: Borrow<Q>,
@@ -159,30 +171,35 @@ where
 
     /// Adds a value to the [`Set`], replacing the existing value, if any, that is equal to the given
     /// one. Returns the replaced value.
+    #[inline]
     pub fn replace(&mut self, value: T) -> Option<T> {
         self.inner.replace(value)
     }
 
     /// Returns `true` if `self` has no elements in common with `other`.
     /// This is equivalent to checking for an empty intersection.
+    #[inline]
     pub fn is_disjoint(&self, other: &Self) -> bool {
         self.inner.is_disjoint(&other.inner)
     }
 
     /// Returns `true` if the [`Set`] is a subset of another,
     /// i.e., `other` contains at least all the values in `self`.
+    #[inline]
     pub fn is_subset(&self, other: &Self) -> bool {
         self.inner.is_subset(&other.inner)
     }
 
     /// Returns `true` if the [`Set`] is a superset of another,
     /// i.e., `self` contains at least all the values in `other`.
+    #[inline]
     pub fn is_superset(&self, other: &Self) -> bool {
         self.inner.is_superset(&other.inner)
     }
 
     /// Visits the values representing the difference,
     /// i.e., the values that are in `self` but not in `other`.
+    #[inline]
     pub fn difference<'a>(&'a self, other: &'a Self) -> Difference<'a, T> {
         Difference {
             inner: self.inner.difference(&other.inner),
@@ -191,6 +208,7 @@ where
 
     /// Visits the values representing the symmetric difference,
     /// i.e., the values that are in `self` or in `other` but not in both.
+    #[inline]
     pub fn symmetric_difference<'a>(&'a self, other: &'a Self) -> SymmetricDifference<'a, T> {
         SymmetricDifference {
             inner: self.inner.symmetric_difference(&other.inner),
@@ -205,6 +223,7 @@ where
     /// one or the other. This can be relevant if `T` contains fields which
     /// are not compared by its `Eq` implementation, and may hold different
     /// value between the two equal copies of `T` in the two sets.
+    #[inline]
     pub fn intersection<'a>(&'a self, other: &'a Self) -> Intersection<'a, T> {
         Intersection {
             inner: self.inner.intersection(&other.inner),
@@ -213,6 +232,7 @@ where
 
     /// Visits the values representing the union,
     /// i.e., all the values in `self` or `other`, without duplicates.
+    #[inline]
     pub fn union<'a>(&'a self, other: &'a Self) -> Union<'a, T> {
         Union {
             inner: self.inner.union(&other.inner),
@@ -224,6 +244,7 @@ impl<T> PartialEq for Set<T>
 where
     T: Eq + Hash,
 {
+    #[inline]
     fn eq(&self, other: &Self) -> bool {
         self.inner == other.inner
     }
@@ -235,6 +256,7 @@ impl<T> FromIterator<T> for Set<T>
 where
     T: Hash + Eq + Ord,
 {
+    #[inline]
     fn from_iter<I>(iter: I) -> Self
     where
         I: IntoIterator<Item = T>,
@@ -249,6 +271,7 @@ impl<'a, T> IntoIterator for &'a Set<T> {
     type Item = &'a T;
     type IntoIter = Iter<'a, T>;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         self.iter()
     }
@@ -258,6 +281,7 @@ impl<'a, T> Extend<&'a T> for Set<T>
 where
     T: Hash + Eq + Ord + Copy + 'a,
 {
+    #[inline]
     fn extend<Iter: IntoIterator<Item = &'a T>>(&mut self, iter: Iter) {
         self.inner.extend(iter)
     }
@@ -267,6 +291,7 @@ impl<T> Extend<T> for Set<T>
 where
     T: Hash + Eq + Ord,
 {
+    #[inline]
     fn extend<Iter: IntoIterator<Item = T>>(&mut self, iter: Iter) {
         self.inner.extend(iter)
     }
@@ -278,6 +303,7 @@ where
 {
     type Output = Set<T>;
 
+    #[inline]
     fn bitand(self, rhs: Self) -> Set<T> {
         Set {
             inner: BitAnd::bitand(&self.inner, &rhs.inner),
@@ -291,6 +317,7 @@ where
 {
     type Output = Set<T>;
 
+    #[inline]
     fn bitor(self, rhs: Self) -> Set<T> {
         Set {
             inner: BitOr::bitor(&self.inner, &rhs.inner),
@@ -304,6 +331,7 @@ where
 {
     type Output = Set<T>;
 
+    #[inline]
     fn bitxor(self, rhs: Self) -> Set<T> {
         Set {
             inner: BitXor::bitxor(&self.inner, &rhs.inner),
@@ -317,6 +345,7 @@ where
 {
     type Output = Set<T>;
 
+    #[inline]
     fn sub(self, rhs: Self) -> Set<T> {
         Set {
             inner: Sub::sub(&self.inner, &rhs.inner),
@@ -333,16 +362,19 @@ pub struct Iter<'a, T> {
 impl<'a, T: 'a> Iterator for Iter<'a, T> {
     type Item = &'a T;
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 }
 
 impl<'a, T: 'a> ExactSizeIterator for Iter<'a, T> {
+    #[inline]
     fn len(&self) -> usize {
         self.inner.len()
     }
@@ -354,6 +386,7 @@ impl<T> IntoIterator for Set<T> {
     type Item = T;
     type IntoIter = IntoIter<T>;
 
+    #[inline]
     fn into_iter(self) -> Self::IntoIter {
         IntoIter {
             inner: self.inner.into_iter(),
@@ -370,16 +403,19 @@ pub struct IntoIter<T> {
 impl<T> Iterator for IntoIter<T> {
     type Item = T;
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 }
 
 impl<T> ExactSizeIterator for IntoIter<T> {
+    #[inline]
     fn len(&self) -> usize {
         self.inner.len()
     }
@@ -407,6 +443,7 @@ where
 }
 
 impl<T> Clone for Difference<'_, T> {
+    #[inline]
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
@@ -420,10 +457,12 @@ where
 {
     type Item = &'a T;
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }
@@ -456,6 +495,7 @@ where
 }
 
 impl<T> Clone for Intersection<'_, T> {
+    #[inline]
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
@@ -469,10 +509,12 @@ where
 {
     type Item = &'a T;
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }
@@ -505,6 +547,7 @@ where
 }
 
 impl<T> Clone for SymmetricDifference<'_, T> {
+    #[inline]
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
@@ -518,10 +561,12 @@ where
 {
     type Item = &'a T;
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }
@@ -554,6 +599,7 @@ where
 }
 
 impl<T> Clone for Union<'_, T> {
+    #[inline]
     fn clone(&self) -> Self {
         Self {
             inner: self.inner.clone(),
@@ -567,10 +613,12 @@ where
 {
     type Item = &'a T;
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.inner.next()
     }
 
+    #[inline]
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.inner.size_hint()
     }