… crates/wasmparser/src/collections/map.rs