I applied `#[inline]` to all trivial methods of types in the `wasmparser::collections` submodule [as stated here](https://github.com/bytecodealliance/wasm-tools/pull/1521#issuecomment-2106164632).

I ran benchmarks on `main` and the PR and built `main` and PR after a `cargo clean` to observe the performance impact of the changes on both compilation and execution. The tl;dr is that compile times are the same, and I only see minor improvements on execution.

Build times:

- `main`: 16.08s user 0.46s system 343% cpu 4.822 total
- PR: 16.04s user 0.46s system 348% cpu 4.736 total

Benchmarks: (PR run after `main` run)

## Default Features

![image](https://github.com/bytecodealliance/wasm-tools/assets/8193155/a61af3cc-da31-4d0a-a1df-d2dbad6567ee)

## Features: `no-hash-maps`

![image](https://github.com/bytecodealliance/wasm-tools/assets/8193155/4dedec89-4577-4a0c-818f-c41de8944fe2)